+++ src/species.rs
@@ -7,40 +7,40 @@ pub struct Species {
     name: String,
     type_1: Type,
     type_2: Type,
-    hp: i32,
-    atk: i32,
-    def: i32,
-    spa: i32,
-    spd: i32,
-    spe: i32,
+    hp: u8,
+    atk: u8,
+    def: u8,
+    spa: u8,
+    spd: u8,
+    spe: u8,
     ability_1: String,
     ability_2: String,
     ability_h: String,
     genderless: bool,
-    m_to_f_ratio: i32,
+    m_to_f_ratio: u8,
     mega: bool,
     gmax: bool,
-    weight: f32,
+    weight: f64,
 }
 impl Species {
     pub fn new(
         name: String,
         type_1: Type,
         type_2: Type,
-        hp: i32,
-        atk: i32,
-        def: i32,
-        spa: i32,
-        spd: i32,
-        spe: i32,
+        hp: u8,
+        atk: u8,
+        def: u8,
+        spa: u8,
+        spd: u8,
+        spe: u8,
         ability_1: String,
         ability_2: String,
         ability_h: String,
         genderless: bool,
-        m_to_f_ratio: i32,
+        m_to_f_ratio: u8,
         mega: bool,
         gmax: bool,
-        weight: f32,
+        weight: f64,
     ) -> Species {
         Species {
             name,
@@ -66,25 +66,25 @@ impl Species {
     pub fn get_name(&self) -> &str {
         &self.name
     }
-    pub fn get_hp(&self) -> i32 {
+    pub fn get_hp(&self) -> u8 {
         self.hp
     }
-    pub fn get_atk(&self) -> i32 {
+    pub fn get_atk(&self) -> u8 {
         self.atk
     }
-    pub fn get_def(&self) -> i32 {
+    pub fn get_def(&self) -> u8 {
         self.def
     }
-    pub fn get_spa(&self) -> i32 {
+    pub fn get_spa(&self) -> u8 {
         self.spa
     }
-    pub fn get_spd(&self) -> i32 {
+    pub fn get_spd(&self) -> u8 {
         self.spd
     }
-    pub fn get_spe(&self) -> i32 {
+    pub fn get_spe(&self) -> u8 {
         self.spe
     }
-    pub fn get_m_to_f_ratio(&self) -> i32 {
+    pub fn get_m_to_f_ratio(&self) -> u8 {
         self.m_to_f_ratio
     }
     pub fn get_genderless(&self) -> bool {
